requirements.txt
pandas>=1.2.0
python-dotenv
requests>=2.18.4
petpy>=2.3.0
psycopg2>=2.8.0
boto3>=1.18
